+++ CHANGELOG.md
@@ -11,6 +11,8 @@ All notable changes to this project will be documented in this file.
 
 ### Chanegd
 
+- Hint section now shows the hint importance.
+- Hint section format in Spoiler log.
 - Default hint distribution for crosskeys preset.
 - Logic now accounts for appearing at the well's bottom in Kakariko when coming out of it.
 - Logic now accounts for appearing underwater in Zora's Domain when coming from the shortcut.